+++ apps/server/src/@types/data/class-defs.ts
@@ -1,5 +1,6 @@
+import type { ClassDef } from "../structure";
 import { prefabModel } from "./def/world/prefab";
 
-export const ClassDefinitions = {
+export const ClassDefinitions: Record<string, ClassDef> = {
   prefab_model: prefabModel
 }